git-svn-id: http://192.168.20.11/sequoiadb/trunk@17199 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/pmd/pmdOptionsMgr.cpp
@@ -1420,7 +1420,7 @@ namespace engine
       rdxUInt( pEX, PMD_OPTION_WEIGHT, weight,
               FALSE, TRUE, 10, FALSE ) ; 
       rdvMinMax( pEX, weight, 0, 100, FALSE ) ;
-      _weight = weight ;
+      _weight = ( UINT8 )weight ;
       // end map
 
       return getResult () ;
